moved IntegrationAPIV2 client to separate folder
