+++ src/sxapi/cli/cli_user.py
@@ -2,7 +2,7 @@ import os
 
 import keyring
 
-from sxapi.base import IntegrationAPIV2
+from sxapi.integrationV2 import IntegrationAPIV2
 from sxapi.publicV2 import PublicAPIV2
